+++ frontend/plugin/statistics/gtkmm/src/StatisticsDialog.cc
@@ -383,7 +383,7 @@ StatisticsDialog::display_statistics(IStatistics::DailyStats *stats)
     }
 
 
-  uint64_t value = stats->misc_stats[IStatistics::STATS_VALUE_TOTAL_ACTIVE_TIME];
+  int64_t value = stats->misc_stats[IStatistics::STATS_VALUE_TOTAL_ACTIVE_TIME];
   daily_usage_label->set_text(Text::time_to_string(value));
 
   // Put the breaks in table.
@@ -589,7 +589,7 @@ StatisticsDialog::on_timer()
 
 
 void
-StatisticsDialog::stream_distance(stringstream &stream, int pixels)
+StatisticsDialog::stream_distance(stringstream &stream, int64_t pixels)
 {
   char buf[64];
